Rename 'components/UntrackedRateMixedChart.vue' to 'components/index/CardsMonitoring/UntrackedRate/Chart.vue'
